first commit
This project involves the development of a spam email detection system utilizing machine learning. By applying the Random Forest algorithm to the Spambase dataset from the Kaggle, the system classifies emails as spam or non-spam (ham). The data is processed with TF-IDF vectorization to extract meaningful features, and the model is optimized using hyperparameter tuning. Achieving high accuracy and insightful feature importance analysis, this project highlights the potential of machine learning in detecting spam effectively.